// interfaces/documents.ts
export interface UserData {
  _id: string;
  name: string;
  email: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
  icon?: string;
  phone?: string;
  language?: string;
  gender?: string;
  auth?: string;
  nickName?: string;
  status?: string;
  assistant?: string;
//  subscription?: string;
}